e_ductus/pyproject.toml
[tool.poetry]
name = "e_ductus"
version = "0.1.0"
description = ""
authors = ["Marius-prog <msabaliauskas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
django-stubs = "^1.7.0"
pytest-django = "^4.1.0"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
